Changes made February 5th 2016
The following classes now implement Serializable:
AbstractUser
Calendar
Job
Park

By association, the following classes now also implement Serializable:
ParkManager
UrbanParksStaffMember
Volunteer

Renamed "UrbanParksDriver.java" to "PrimaryDriver.java"
This was because I thought it might be confusing when we implemented
UrbanParksStaffMemberDriver later.

main in PrimaryDriever now:
Reads in the list of 'all' users from a .txt file
This file is written in the following format:
userEmail@domain.com,firstName,lastName,userType
'All' users is currently defined as 1. It should keep reading as long as
there is more of the file to read.

Committed a crime against nature with line 50.

Once the file has been read in, the user is prompted to log in
This is currently a hardcoded String.

Once the user 'enters' their email, they are compared against some if
statements about what type of user they are

I'll work on PMDriver itself next time I get a chance to code.