Changes February 8th
Park is now properly Serializable
Removed serialVersionID from Park
The class is going to update again before that becomes useful, so it's
removed for now as a reminder for when the class is done.

Job is now properly Serializable
Job now contains the stub for the equals(Object o) method.