Refactoring
Added new kinds of exceptions
-CalendarFullExceptions are thrown when the upcomingJobCollection's size
exceeds the MAX_JOBS constant.
-CalendarWeekFullExceptions are throw when an added job would cause too
many jobs to be scheduled that week
-JobTooLongExceptions are thrown when a Job is longer than 48 hours in
length
-JobTimeTravelExceptions are thrown when a job ends before it begins.
-JobToThePastExceptions are thrown when a job has already occured and is
tried to be put in the Calendar.
-JobToTheFutureExceptions are thrown when a job occurs more than 90 days
in the future.
-AlreadyVolunteeredExceptions occur when a volunteer already has that
job in their jobs list
-ConflictingJobCommitmentException occurs when a Volunteer has already
volunteered for a job that day
-JobIsFullException occurs when there are no more Volunteers needed and
another Volunteer attempts to volunteer.

Added method public boolean editable to jobs.
-it reads: 'return volunteers.size() ==0;
--You can't edit a job that already has volunteers.

Intentionally gutted a couple of the drivers
-The methods around where new exceptions were thrown weren't going to
work anyway

Moved code around in PrimaryDriver so it'll be a little easier to read

Removed some validation logic away from the ParkManagerDriver class