Editting changes
Job.java
Implemented a copy constructor to support job editting.

Park.java
Implemented equals(Object o)
-It was needed for a function in ParkManagerDriver

ParkManagerDriver
Made it a little more obvious who (and what) is logged in
Took out UrbanParkCalendar as a parameter for most methods
-We can use the field instead.
Changed how the driver creates a job before attempting to add it to the
Calendar.
When selecting a park, only the name is displayed.
Changed the logic slightly for displaying jobs in editJob.
-Now it gets the master list and immediatly removes all jobs not from
that park instead of saving it for later.
-This functionality should be moved to a method

PrimaryDriver
Added a line which prints all acceptable login information (for
debugging)

UrbanParkCalendar
Removed SerialUID
Noted why jobsToCheck in the method updateCalendar() exists
Uncommented the check that throws DuplicateJobExistsException
Changed how editJobTitle and editJobDesc work
-They now copy jobs (Except the volunteers!), edit the field that needed
it, then replaces the job in the calendar.

UrbanParkCalendarTest.java
Updated some tests I noticed errors in.

Remaining issues:

submitNewJob should be broken up
-It's over 100 lines long

We should move to the global choice design pattern as used in the other
two user Drivers.
-This will require more error checking, but can be put into single
methods (Probably in SharedUserDriverFunctions)

Displaying all jobs in a park should be moved to a new method.
-This would include validation

When attempting to edit multiple fields for a job at the same time, new
copies of it emerge.
-I'm pretty sure I know what's causing this and it'll be the first thing
I fix.