Additional changes
PrimaryDriver.java
Now has a "debug" command that prints the count of jobs in the upcoming
job list, the amount of Volunteers (specifically Volunteers) that exist
in the Calendar and the emails of all the users who could log in
The "reset" command is a little more explicit now
Renamed 'fabricateInformation' to 'resetInformation'

SharedUserDriverFunctions.java
printAllUpcomingJobs now starts printing jobs starting at 1.

UrbanParkStaffMemberDriver.java
Changed the login message

VolunteerDriver.java
Changed the logic of menu navigation to use integers instead of Strings
Changed the login message
Changed some Strings that were prompting for user input.
Implemented and fixed a bug that was causing the program to end if you
signed up for difficult work