More improvements
Changed all the tree jobs so it was more obvious they weren't the same
job.
Made defeating Tenenberg and Bigfoot harder jobs
Added a loop which adds a bunch of 'loitering' jobs starting three weeks
from today.
Changed the displayed String when attempting to volunteer for a job that
was full.
Moved the "You can volunteer again" print out
Added > in front of prompts to make where you were in the program more
obvious